MdeModulePkg/PeiCore: avoid EFI_IMAGE_MACHINE_TYPE_SUPPORTED to check arch
The EFI_IMAGE_MACHINE_TYPE_SUPPORTED() macro is intended for deciding
whether a certain PE/COFF machine type is compatible with the current
system. This has nothing to do with the architecture we are compiling
for, especially in cases where PE/COFF images built for X64 may be
executed in a VM on foreign architectures.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>