MdeModulePkg/AcpiTableDxe: inhibit table publication until we have a FADT
Since commit 78c41ff519b1 ("ArmVirtPkg/FdtClientDxe: make DT table
installation !ACPI dependent"), we inhibit the installation of the core
ACPI tables when booting in DT mode, to relieve the OS from having to
decide which hardware description to use. However, as it turns out,
the presence of the RamdiskDxe or BGRT drivers, some ACPI tables are
still registered with the protocol, and published under the ACPI entry
point configuration table, ignoring the fact that there is no way the
OS can boot with only NFIT and BGRT tables present.

So let's only publish the ACPI tables if we can reasonably assume that
the tables that the OS requires have been registered with the protocol,
by checking that we have either FADT1 or FADT3 table (or both) before
installing the configuration tables.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>